+++ StereoKitC/StereoKitC.vcxproj
@@ -186,6 +186,7 @@ xcopy "$(ProjectDir)stereokit_ui.h" "$(ProjectDir)..\bin\distribute\include\" /d
     <ClCompile Include="tools\file_picker.cpp" />
     <ClCompile Include="tools\tools.cpp" />
     <ClCompile Include="tools\virtual_keyboard.cpp" />
+    <ClCompile Include="ui\interactors.cpp" />
     <ClCompile Include="ui\stereokit_ui.cpp" />
     <ClCompile Include="ui\ui_core.cpp" />
     <ClCompile Include="ui\ui_layout.cpp" />
@@ -299,6 +300,7 @@ xcopy "$(ProjectDir)stereokit_ui.h" "$(ProjectDir)..\bin\distribute\include\" /d
     <ClInclude Include="systems\world.h" />
     <ClInclude Include="tools\tools.h" />
     <ClInclude Include="tools\virtual_keyboard.h" />
+    <ClInclude Include="ui\interactors.h" />
     <ClInclude Include="ui\ui_core.h" />
     <ClInclude Include="ui\ui_layout.h" />
     <ClInclude Include="ui\ui_theming.h" />
